splits cluster.js in 2 files
