+++ test/cluster.test.js
@@ -3,9 +3,9 @@
 require('./test-setup.js');
 
 const
+  cleanrequire = require('@everymundo/cleanrequire'),
   sinon    = require('sinon'),
   {expect} = require('chai');
-  // cleanrequire = require('@everymundo/cleanrequire'),
   // clone    = arg => JSON.parse(JSON.stringify(arg));
 
 describe('cluster.js', () => {
@@ -30,224 +30,45 @@ describe('cluster.js', () => {
 
   context('on load', () => {
     it('should export expected functions', () => {
-      expect(server.forkAWorker).to.be.instanceof(Function);
-      expect(server.createWorkers).to.be.instanceof(Function);
-      expect(server.configKillSignals).to.be.instanceof(Function);
-      expect(server.configClusterEvents).to.be.instanceof(Function);
-      expect(server.savePidFile).to.be.instanceof(Function);
-      expect(server.registerToEureka).to.be.instanceof(Function);
       expect(server.initMaster).to.be.instanceof(Function);
       expect(server.initWorker).to.be.instanceof(Function);
       expect(server.init).to.be.instanceof(Function);
     });
   });
 
-  describe('forkAWorker', () => {
-    const { forkAWorker } = server;
+  describe('initMaster', () => {
+    const clusterFeaturesLib = require('../cluster-features');
+    const clusterFeaturesFunctionNames = [
+      'savePidFile',
+      'createWorkers',
+      'configClusterEvents',
+      'configKillSignals',
+      'registerToEureka',
+    ];
 
-    it('should call cluster.fork', () => {
-      const cluster = {fork: sinon.spy(() => ({send: () => {}, id: 100}))};
-
-      forkAWorker(cluster);
-      expect(cluster.fork).to.have.property('calledOnce', true);
-    });
-
-    it('should call worker.send', () => {
-      const send = sinon.spy((obj) => {
-        expect(obj).to.have.property('type', 'setWorkerId');
-        expect(obj).to.have.property('workerId', 100);
-      });
-
-      const cluster = {fork: sinon.spy(() => ({send, id: 100}))};
-
-      forkAWorker(cluster);
-      expect(cluster.fork).to.have.property('calledOnce', true);
-      expect(send).to.have.property('calledOnce', true);
-    });
-  });
-
-  describe('createWorkers', () => {
-    const { createWorkers } = server;
-    const os = require('os');
-    const { config } = require('../config');
+    let result;
 
     beforeEach(() => {
-      box.stub(os, 'cpus').callsFake(() => ({length: 2}));
+      result = [];
+      clusterFeaturesFunctionNames.forEach(func =>
+        box.stub(clusterFeaturesLib, func).callsFake(() => { result.push(func); }));
     });
 
-    context('when NUM_OF_WORKERS is set', () => {
-      beforeEach(() => {
-        box.stub(config, 'NUM_OF_WORKERS').value(1);
-      });
+    it('should call cluster features functions in order passing the cluster arg', () => {
+      const { initMaster } = cleanrequire('../cluster');
+      const cluster = {};
+      initMaster(cluster);
 
-      it('should not call os.cpus()', () => {
-        const send = sinon.spy((obj) => {
-          expect(obj).to.have.property('type', 'setWorkerId');
-          expect(obj).to.have.property('workerId', 100);
-        });
+      expect(result).to.deep.equal(clusterFeaturesFunctionNames);
 
-        const cluster = {
-          on:   sinon.spy((msg, cb) => cb()),
-          fork: sinon.spy(() => ({ send, id: 100 })),
-        };
-
-        createWorkers(cluster);
-
-        expect(os.cpus).to.have.property('calledOnce', false);
-        expect(cluster.fork).to.have.property('calledOnce', true);
-        expect(send).to.have.property('calledOnce', true);
-      });
-    });
-
-    context('when NUM_OF_WORKERS is NOT set', () => {
-      beforeEach(() => {
-        box.stub(config, 'NUM_OF_WORKERS').value(0);
-      });
-
-      it('should not call os.cpus()', () => {
-        const send = sinon.spy((obj) => {
-          expect(obj).to.have.property('type', 'setWorkerId');
-          expect(obj).to.have.property('workerId', 100);
-        });
-
-        const cluster = {
-          on:   sinon.spy((msg, cb) => cb()),
-          fork: sinon.spy(() => ({ send, id: 100 })),
-        };
-
-        createWorkers(cluster);
-
-        expect(cluster.fork).to.have.property('calledTwice', true);
-        expect(send).to.have.property('calledTwice', true);
-        expect(os.cpus).to.have.property('calledOnce', true);
-      });
-    });
-  });
-
-  describe('configKillSignals', () => {
-    const { configKillSignals } = server;
-
-    beforeEach(() => {
-      box.stub(process, 'on').callsFake((msg, cb) => cb());
-    });
-
-    it('should call process.on', (done) => {
-      const send = sinon.spy((obj) => {
-        expect(obj).to.have.property('type', 'stop');
-        done();
-      });
-
-      const cluster = {
-        workers: { 1: {}, 2: { send }},
-      };
-
-      configKillSignals(cluster);
-
-      expect(process.on).to.have.property('calledOnce', true);
-    });
-  });
-
-  describe('configClusterEvents', () => {
-    const { configClusterEvents } = server;
-    const worker = {process:{pid: 1}};
-    const code   = 0;
-    const signal = 'SIGNAL';
-
-    let clock;
-    beforeEach(() => { clock = sinon.useFakeTimers(); });
-
-    afterEach(() => { clock.restore(); });
-
-    context('with signal', () => {
-      it('should call cluster.on', () => {
-        const send = sinon.spy((obj) => {
-          expect(obj).to.have.property('type', 'stop');
-        });
-
-        const cluster = {
-          workers: { 1: {}, 2: { send } },
-          on: sinon.spy((msg, cb) => cb(worker, code, signal)),
-        };
-
-        configClusterEvents(cluster);
-
-        expect(cluster.on).to.have.property('calledOnce', true);
-      });
-    });
-
-    context('WITHOUT signal', () => {
-      it('should call cluster.on', () => {
-        const send = sinon.spy((obj) => {
-          expect(obj).to.have.property('type', 'stop');
-        });
-
-        const cluster = {
-          workers: { 1: {}, 2: { send } },
-          on: sinon.spy((msg, cb) => cb(worker, code)),
-        };
-
-        configClusterEvents(cluster);
-
-        expect(cluster.on).to.have.property('calledOnce', true);
-      });
-    });
-  });
-
-  describe('savePidFile', () => {
-    const fs = require('fs');
-
-    beforeEach(() => {
-      box.stub(fs, 'writeFileSync').callsFake(() => {});
-    });
-
-    it('should call fs.writeFileSync', () => {
-      const { savePidFile } = server;
-      savePidFile();
-      expect(fs.writeFileSync).to.have.property('calledOnce', true);
-    });
-  });
-
-  describe('registerToEureka', () => {
-    const emEurekaLib = require('@everymundo/em-eureka');
-
-    let calledWithArgs;
-
-    context('When asyncClientFromConfigService resolves', () => {
-      beforeEach(() => {
-        calledWithArgs = undefined;
-        box.stub(emEurekaLib, 'asyncClientFromConfigService')
-          .callsFake(args => new Promise((resolve) => { calledWithArgs = args; resolve(); }));
-      });
-
-      it('should call asyncClientFromConfigService and resolve', () => {
-        const { registerToEureka } = server;
-
-        return registerToEureka().then(() => {
-          expect(emEurekaLib.asyncClientFromConfigService).to.have.property('calledOnce', true);
-          expect(calledWithArgs).to.have.property('port');
-          expect(calledWithArgs).to.have.property('securePort');
-        });
-      });
-    });
-
-    context('When asyncClientFromConfigService rejects', () => {
-      const error = new Error('asyncClientFromConfigService error');
-
-      beforeEach(() => {
-        calledWithArgs = undefined;
-        box.stub(emEurekaLib, 'asyncClientFromConfigService')
-          .callsFake(args => new Promise((_, reject) => { calledWithArgs = args; reject(error); }));
-      });
-
-      it('should call asyncClientFromConfigService and REJECT', () => {
-        const { registerToEureka } = server;
-
-        return registerToEureka().catch(() => {
-          expect(emEurekaLib.asyncClientFromConfigService).to.have.property('calledOnce', true);
-          expect(calledWithArgs).to.have.property('port');
-          expect(calledWithArgs).to.have.property('securePort');
-        });
-      });
+      [
+        'createWorkers',
+        'configClusterEvents',
+        'configKillSignals',
+        'registerToEureka',
+      ].forEach(func =>
+        expect(clusterFeaturesLib[func].calledWith(cluster))
+          .to.equal(true, `${func} did not receive correct cluster argument`));
     });
   });
 
@@ -267,7 +88,29 @@ describe('cluster.js', () => {
 
   describe('init', () => {
     context('cluster.isMaster is TRUE', () => {
-      it('should test its functionality');
+      const clusterFeaturesLib = require('../cluster-features');
+      const clusterFeaturesFunctionNames = [
+        'savePidFile',
+        'createWorkers',
+        'configClusterEvents',
+        'configKillSignals',
+        'registerToEureka',
+      ];
+
+      let result;
+
+      beforeEach(() => {
+        result = [];
+        clusterFeaturesFunctionNames.forEach(func =>
+          box.stub(clusterFeaturesLib, func).callsFake(() => { result.push(func); }));
+      });
+
+      it('should call initMaster', () => {
+        const { init } = cleanrequire('../cluster');
+        init({ isMaster: true });
+
+        expect(result).to.deep.equal(clusterFeaturesFunctionNames);
+      });
     });
 
     context('cluster.isMaster is FALSE', () => {
@@ -278,7 +121,7 @@ describe('cluster.js', () => {
         box.stub(worker, 'init').callsFake(() => { });
       });
 
-      it('should call server.init()', () => {
+      it('should call initWorker()', () => {
         init({isMaster: false});
         expect(worker.init).to.have.property('calledOnce', true);
       });
